First iteration of relaxing the grid and seeding the grid
The grid is relaxed by finding the closest square with the same center of mass and a predefined side length and using that to apply a force to the grid points towards the perfect square.

Added a seed to the grid generator.

The indices of the quads are no longer shown in editor.