Correct mesh selection and rotation
The mesh is correctly chosen used on a key based on the marching cubes corners.
The rotation is determined by the marching cubes corners the tile has compared to the "default tile".
The rotated mesh is done by rotating the cage.
Replaced meshes to have correct "default" rotation.

Mesh deformation is still off with the middle vertices.